RUB-1127: widen the fuzz-crate fee consumers to u128
The fuzz crate is excluded from the Rust workspace (clients/rust/Cargo.toml
exclude=["fuzz"]), so cargo test --workspace never compiled it while the fee
scalar widened to u128. Two targets stopped building: tx_relay_announce builds
RelayTxMetadata.fee and validate_tx_local builds PrecomputedTxContext.fee.

Both derive the fuzzed value exactly as before from the same input bytes and
only widen at the boundary; what is fuzzed is unchanged.

+++ clients/rust/fuzz/fuzz_targets/validate_tx_local.rs
@@ -200,7 +200,7 @@ fuzz_target!(|data: &[u8]| {
     // Sum input values for fee calculation.
     let sum_in: u64 = resolved_inputs.iter().map(|e| e.value).fold(0u64, |a, b| a.saturating_add(b));
     let sum_out: u64 = tx.outputs.iter().map(|o| o.value).fold(0u64, |a, b| a.saturating_add(b));
-    let fee = sum_in.saturating_sub(sum_out);
+    let fee = u128::from(sum_in.saturating_sub(sum_out));
 
     let ptc = rubin_consensus::PrecomputedTxContext {
         tx_index: 1,
